refactor(liquidation): extract calculateDistributionPlan
